# Conflicts:
#	.idea/file.template.settings.xml
#	.idea/modules.xml
#	.idea/vcs.xml

+++ engine/framework/include/2XD2/framework/resource_manager/Animation.h
@@ -1,3 +1,7 @@
+// Copyright (c) 2026 Rune Suy and the 2XD2-Engine contributors.
+// Licensed under the MIT License.
+//
+
 //
 // Created by rune-suy on 12/17/25.
 //
@@ -21,9 +25,9 @@ namespace e2XD::framework
 
     public:
         Animation() = default;
-        Animation(const std::initializer_list<const sf::Texture*>& frameList, float frameDuration,
+        Animation(const std::initializer_list<const sf::Texture*>& frameList, double frameDuration,
                   bool isLooping = false);
-        Animation(const std::vector<const sf::Texture*>& frameList, float frameDuration, bool isLooping = false);
+        Animation(const std::vector<const sf::Texture*>& frameList, double frameDuration, bool isLooping = false);
 
         /**
          * Add a frame to the animation.
@@ -35,7 +39,7 @@ namespace e2XD::framework
          * Set the duration of each frame in seconds.
          * @param duration Duration in seconds.
          */
-        void setFrameDuration(float duration);
+        void setFrameDuration(double duration);
 
         /**
          * Set whether the animation should loop.
@@ -53,7 +57,7 @@ namespace e2XD::framework
          *
          * @return the frame duration in seconds.
          */
-        [[nodiscard]] float getFrameDuration() const;
+        [[nodiscard]] double getFrameDuration() const;
 
         /**
          *
